1. 下载文件（包括位图）进度通知监听，即设置HttpListener 2. request可设置id，以唯一标记一个请求；request可以设置tag，以来保存一些数据。

+++ library/src/com/litesuits/http/parser/InputStreamParser.java
@@ -13,7 +13,7 @@ import java.io.InputStream;
 public abstract class InputStreamParser extends DataParser<InputStream> {
 
 	@Override
-	public InputStream parseData(InputStream stream, int len, String charSet) throws IOException {
+	public InputStream parseData(InputStream stream, long len, String charSet) throws IOException {
 		return stream;
 	}
